refactor(sharing): extract transport layer into veld-share crate
Moves the share protocol's transport layer out of veld-daemon into a new
library crate so veld-daemon and the upcoming veld-gateway compile
against the same implementation and cannot drift:

- proto (control/open-stream frames), forward::splice, join (dial +
  forward_local), host (HostShare, accept/deny/serve), endpoint
  (ALPN, RelayChoice, bind_endpoint, relay token resolution, node key
  persistence) move verbatim to crates/veld-share
- relay auth-denial detection (RelayAuth, relay_auth_status,
  is_relay_auth_denial) moves from the daemon's manager into
  veld-share::endpoint so any joiner can use it
- resolve_secret and resolve_embedded_tokens become pub: the gateway
  reuses the same SecretSource resolution semantics
- iroh/iroh-tickets become workspace dependencies so every crate pins
  the same version
- veld-gateway crate stubbed as a workspace member (implemented next)

Pure refactor: no behavior change; all tests pass unchanged.
