+++ Cargo.toml
@@ -4,6 +4,8 @@ members = [
     "crates/veld",
     "crates/veld-helper",
     "crates/veld-daemon",
+    "crates/veld-share",
+    "crates/veld-gateway",
 ]
 resolver = "2"
 
@@ -15,6 +17,9 @@ rust-version = "1.85"
 
 [workspace.dependencies]
 veld-core = { path = "crates/veld-core" }
+veld-share = { path = "crates/veld-share" }
+iroh = "1"
+iroh-tickets = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 tokio = { version = "1", features = ["full"] }
