+++ crates/veld-gateway/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "veld-daemon"
+name = "veld-gateway"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 
 [[bin]]
-name = "veld-daemon"
+name = "veld-gateway"
 path = "src/main.rs"
 
 [dependencies]
@@ -16,17 +16,7 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 anyhow = { workspace = true }
-thiserror = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
-chrono = { workspace = true }
-dirs = { workspace = true }
-libc = "0.2"
-axum = "0.8"
-reqwest = { workspace = true }
-uuid = { workspace = true }
 iroh = { workspace = true }
 iroh-tickets = { workspace = true }
-
-[dev-dependencies]
-tempfile = "3"
