+++ crates/veld-gateway/src/main.rs
@@ -1,0 +1,5 @@
+//! veld-gateway — the public web gateway for Veld sharing (SHARING_V2.md §5).
+
+fn main() {
+    println!("veld-gateway: not yet implemented");
+}
