+++ crates/veld-share/Cargo.toml
@@ -1,32 +1,21 @@
 [package]
-name = "veld-daemon"
+name = "veld-share"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 
-[[bin]]
-name = "veld-daemon"
-path = "src/main.rs"
-
 [dependencies]
 veld-core = { workspace = true }
-veld-share = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 anyhow = { workspace = true }
-thiserror = { workspace = true }
 tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-chrono = { workspace = true }
 dirs = { workspace = true }
-libc = "0.2"
-axum = "0.8"
-reqwest = { workspace = true }
-uuid = { workspace = true }
 iroh = { workspace = true }
 iroh-tickets = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3"
+uuid = { workspace = true }
