+++ crates/veld-share/src/host.rs
