+++ crates/veld-share/src/join.rs
